The size field in msgpack_object_array is a uint32.

+++ lib/ffi/msgpack/msg_array.rb
@@ -5,7 +5,7 @@ module FFI
   module MsgPack
     class MsgArray < FFI::Struct
 
-      layout :size, :size_t,
+      layout :size, :uint32,
              :ptr, :pointer
 
       #
